Extract specs for profiling component to separate file
This move was done with as minimal changes as possible.
